# PERF: contiguity, less gil in join algos