@jreback 2021-06-17
@meeseeksdev backport 1.3.x